cleanup and fix bugs at DAO class in order impl RMI.
