+++ src/main/java/Main.java
@@ -5,7 +5,7 @@ import javafx.application.Application;
 import javafx.stage.Stage;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
-import persistence.DatabaseManager;
+import persistence.CalendarDAOImpl;
 
 /**
  * Name: Mr.Chatchapol Rasameluangon
@@ -14,27 +14,23 @@ import persistence.DatabaseManager;
 
 
 public class Main extends Application {
-    private DatabaseManager dao;
-    ApplicationContext context;
+    private CalendarDAOImpl dao;
+    private ApplicationContext context;
 
 
     @Override
     public void init() throws Exception {
         // Main source initialized, attempt to create event.
         context = new ClassPathXmlApplicationContext("main-context.xml");
-        dao = context.getBean("databaseManager", DatabaseManager.class);
+        dao = context.getBean("calendarDAOImpl", CalendarDAOImpl.class);
         dao.setup();
         dao.getTaskFuture().get();
     }
 
     @Override
     public void start(Stage primaryStage) throws Exception {
-        // model
         EventManager em = context.getBean("eventManager", EventManager.class);
-        em.loadEvent();
-        // view
         ViewManager vm = new ViewManager(primaryStage);
-        // main control
         MainController mc = new MainController(em, vm);
         mc.start();
     }
@@ -42,7 +38,6 @@ public class Main extends Application {
     @Override
     public void stop() {
         dao.close();
-        System.out.println("program exits");
     }
 
     public static void main(String[] args) {
