+++ src/main/java/common/search/Searcher.java
@@ -1,4 +1,4 @@
-package common;
+package common.search;
 
 import javafx.collections.ObservableList;
 import model.Event;
