--- src/main/resources/fxml/MasterView.fxml
@@ -1,124 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<!--
-*Name: Mr.Chatchapol Rasameluangon
-*ID:   5810404901
--->
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.effect.DropShadow?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" prefHeight="400.0" prefWidth="600.0"
-            xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="view.MasterView">
-    <MenuBar layoutY="1.0" prefHeight="25.0" prefWidth="600.0">
-        <Menu mnemonicParsing="false" text="File">
-            <MenuItem mnemonicParsing="false" text="Close"/>
-        </Menu>
-        <Menu mnemonicParsing="false" text="Edit">
-            <MenuItem mnemonicParsing="false" text="Delete"/>
-        </Menu>
-        <Menu mnemonicParsing="false" text="Help">
-            <MenuItem mnemonicParsing="false" text="About"/>
-        </Menu>
-    </MenuBar>
-    <SplitPane dividerPositions="0.43478260869565216" layoutY="24.0" prefHeight="375.0" prefWidth="600.0">
-        <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
-            <GridPane prefHeight="373.0" prefWidth="230.0" AnchorPane.bottomAnchor="0.0"
-                      AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-                <columnConstraints>
-                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0"/>
-                </columnConstraints>
-                <rowConstraints>
-                    <RowConstraints maxHeight="336.0" minHeight="10.0" prefHeight="336.0" vgrow="SOMETIMES"/>
-                    <RowConstraints maxHeight="217.0" minHeight="10.0" prefHeight="37.0" vgrow="SOMETIMES"/>
-                </rowConstraints>
-                <FlowPane alignment="CENTER_RIGHT" prefHeight="200.0" prefWidth="200.0"
-                          GridPane.rowIndex="1">
-                    <Button fx:id="addEventButton" mnemonicParsing="false" onAction="#handleAddEvent"
-                            text="Add Event">
-                        <FlowPane.margin>
-                            <Insets right="10.0"/>
-                        </FlowPane.margin>
-                    </Button>
-                </FlowPane>
-                <TableView fx:id="eventTable" prefHeight="200.0" prefWidth="200.0">
-                    <columns>
-                        <TableColumn fx:id="colorCol" prefWidth="13.0"/>
-                        <TableColumn fx:id="dateCol" prefWidth="106.0" text="Date"/>
-                        <TableColumn fx:id="nameCol" prefWidth="136.0" text="Name"/>
-                    </columns>
-                </TableView>
-            </GridPane>
-        </AnchorPane>
-        <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="265.0" prefWidth="310.0">
-            <GridPane layoutX="8.0" layoutY="83.0" prefHeight="179.0" prefWidth="316.0">
-                <columnConstraints>
-                    <ColumnConstraints hgrow="SOMETIMES" maxWidth="169.0" minWidth="10.0" prefWidth="109.0"/>
-                    <ColumnConstraints hgrow="SOMETIMES" maxWidth="239.0" minWidth="10.0" prefWidth="239.0"/>
-                </columnConstraints>
-                <rowConstraints>
-                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES"/>
-                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES"/>
-                    <RowConstraints maxHeight="40.0" minHeight="10.0" prefHeight="33.0" vgrow="SOMETIMES"/>
-                    <RowConstraints maxHeight="56.0" minHeight="10.0" prefHeight="56.0" vgrow="SOMETIMES"/>
-                </rowConstraints>
-                <Label text="Name" textFill="#686363">
-                    <font>
-                        <Font name="Verdana Bold" size="12.0"/>
-                    </font>
-                </Label>
-                <Label text="Time" textFill="#686363" GridPane.rowIndex="1">
-                    <font>
-                        <Font name="Verdana Bold" size="12.0"/>
-                    </font>
-                </Label>
-                <Label text="Tag" textFill="#686363" GridPane.rowIndex="2">
-                    <font>
-                        <Font name="Verdana Bold" size="12.0"/>
-                    </font>
-                </Label>
-                <Label fx:id="timeLabel" text="?" GridPane.columnIndex="1" GridPane.rowIndex="1">
-                    <font>
-                        <Font name="Verdana" size="12.0"/>
-                    </font>
-                </Label>
-                <Label fx:id="tagLabel" text="?" GridPane.columnIndex="1" GridPane.rowIndex="2">
-                    <font>
-                        <Font name="Verdana" size="12.0"/>
-                    </font>
-                </Label>
-                <Label text="Note" textFill="#686363" GridPane.rowIndex="3">
-                    <font>
-                        <Font name="Verdana Bold" size="12.0"/>
-                    </font>
-                </Label>
-                <TextArea fx:id="noteTextArea" editable="false" prefHeight="200.0" prefWidth="200.0"
-                          style="-fx-background-color: transparent;" wrapText="true" GridPane.columnIndex="1"
-                          GridPane.rowIndex="3"/>
-                <Label fx:id="nameLabel" text="?" GridPane.columnIndex="1">
-                    <font>
-                        <Font name="Verdana" size="12.0"/>
-                    </font>
-                </Label>
-            </GridPane>
-            <Label layoutX="8.0" layoutY="46.0" text="EVENT OVERVIEW" textFill="#5b5757">
-                <font>
-                    <Font name="Verdana Bold" size="20.0"/>
-                </font>
-            </Label>
-            <Button fx:id="removeEventButton" layoutX="240.0" layoutY="342.0" mnemonicParsing="false"
-                    onAction="#handleRemoveEvent" prefHeight="25.0" prefWidth="71.0" text="Remove"/>
-            <Button fx:id="editEventButton" layoutX="187.0" layoutY="342.0" mnemonicParsing="false"
-                    onAction="#handleEditEvent" prefHeight="25.0" prefWidth="46.0" text="Edit"/>
-            <AnchorPane fx:id="colorRectangle" layoutY="1.0" prefHeight="25.0" prefWidth="335.0"
-                        style="-fx-background-color: #FFFFFF;">
-                <effect>
-                    <DropShadow height="9.14" radius="7.035"/>
-                </effect>
-            </AnchorPane>
-        </AnchorPane>
-    </SplitPane>
-</AnchorPane>
